Fixed ampersand issue for offer rules https://github.com/BroadleafCommerce/QA/issues/4663
