+++ core/broadleaf-framework/src/main/java/org/broadleafcommerce/core/offer/service/processor/AbstractBaseProcessor.java
@@ -47,8 +47,6 @@ import org.broadleafcommerce.core.order.service.type.FulfillmentType;
 import org.broadleafcommerce.profile.core.domain.Customer;
 import org.joda.time.DateTimeZone;
 import org.joda.time.LocalDateTime;
-import org.springframework.beans.factory.annotation.Value;
-import javax.annotation.Resource;
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.Collection;
@@ -61,6 +59,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.TimeZone;
+import javax.annotation.Resource;
 
 /**
  * 
@@ -77,9 +76,6 @@ public abstract class AbstractBaseProcessor implements BaseProcessor {
     
     @Resource(name = "blOfferServiceExtensionManager")
     protected OfferServiceExtensionManager extensionManager;
-
-    @Value("${exploitProtection.xssEnabled:false}")
-    protected boolean xssExploitProtectionEnabled;
 
     protected final PromotableOfferUtility promotableOfferUtility;
 
@@ -273,7 +269,7 @@ public abstract class AbstractBaseProcessor implements BaseProcessor {
                 extensionManager.applyAdditionalRuleVariablesForItemOfferEvaluation(orderItem, vars);
             }
 
-            Boolean expressionOutcome = executeExpression(escapeSpecialCharacters(criteria.getMatchRule()), vars);
+            Boolean expressionOutcome = executeExpression(criteria.getMatchRule(), vars);
             if (expressionOutcome != null && expressionOutcome) {
                 appliesToItem = true;
             }
@@ -284,18 +280,6 @@ public abstract class AbstractBaseProcessor implements BaseProcessor {
         return appliesToItem;
     }
 
-    protected String escapeSpecialCharacters(String s) {
-        if (xssExploitProtectionEnabled) {
-            Map<String, String> stringMapping = new HashMap<>();
-            stringMapping.put("&", "&amp;");
-
-            for (Map.Entry<String, String> entry : stringMapping.entrySet()) {
-                s = s.replace(entry.getKey(), entry.getValue());
-            }
-        }
-        return s;
-    }
-    
     /**
      * Private method used by couldOfferApplyToOrder to execute the MVEL expression in the
      * appliesToOrderRules to determine if this offer can be applied.
